why no worky?

+++ Lighting/Src/ws2812.cpp
@@ -15,7 +15,7 @@ static constexpr uint8_t PWM_LO = 19;
 static constexpr uint8_t PWM_HI = 38;
 
 void initialize_bank_output_buffer_off(uint8_t *bank_out_buff, uint8_t num_led, uint8_t num_pad) {
-	uint8_t buffer_size = (num_led + 2*num_pad) * 24;
+	uint16_t buffer_size = (num_led + 2*num_pad) * 24;
 	for (int i = 0; i < buffer_size; ++i) {
 		// if PAD
 		if (i < 24*num_pad || i >= buffer_size - 24*num_pad) {
@@ -27,7 +27,7 @@ void initialize_bank_output_buffer_off(uint8_t *bank_out_buff, uint8_t num_led, 
 }
 
 void initialize_bank_output_buffer_on(uint8_t *bank_out_buff, uint8_t num_led, uint8_t num_pad) {
-	uint8_t buffer_size = (num_led + 2*num_pad) * 24;
+	uint16_t buffer_size = (num_led + 2*num_pad) * 24;
 	for (int i = 0; i < buffer_size; ++i) {
 		// if PAD
 		if (i < 24*num_pad || i >= buffer_size - 24*num_pad) {
